pyproject.toml
[tool.poetry]
name = "virtual-assistant"
version = "0.1.0"
description = "virtual assistant for DS News"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.30.3"
python-telegram-bot = "^21.2"
python-dotenv = "^1.0.1"
gitpython = "^3.1.43"
chardet = "^5.2.0"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.2"
requests-mock = "^1.12.1"
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
